Reorganizing unit tests
