+++ kernel/UnitTests.cpp
@@ -1,6 +1,5 @@
 #include "UnitTests.h"
 
-#include <bitset>
 #include <climits>
 #include <cstdint>
 #include <cstring>
@@ -8,6 +7,8 @@
 #include <utility>
 
 #include "AArch64/CPU.h"
+#include "UnitTests/KernelStdlib/BitsetTests.h"
+#include "UnitTests/Framework.h"
 #include "MemoryManager.h"
 #include "Print.h"
 #include "Utils.h"
@@ -17,279 +18,7 @@
 
 namespace
 {
-    ///////////////////////////////////////////////////////////////////////////
-    // Test "framework"
-    ///////////////////////////////////////////////////////////////////////////
-
-    // #TODO
-    // Make into some sort of terminal output library to control colors, positions, etc.
-
-    /**
-     * Formats a colored string for terminal output
-     * 
-     * @param arBuffer Buffer to output to
-     * @param apString The string to output
-     * @param aColor The color the string should be (terminal escape sequence color)
-     */
-    template<std::size_t BufferSize>
-    void FormatColoredString(char (&arBuffer)[BufferSize], const char* const apString, const uint32_t aColor)
-    {
-        Print::FormatToBuffer(arBuffer, "\x1b[{}m{}\x1b[m", aColor, apString);
-    }
-
-    /**
-     * Tiny helper to emit pass or failure based on a result bool
-     * 
-     * @param aResult The result of the test
-     * @param apMessage The message to emit
-     */
-    void EmitTestResult(const bool aResult, const char* const apMessage)
-    {
-        char passFailMessage[32];
-        if (aResult)
-        {
-            FormatColoredString(passFailMessage, "PASS", 32 /*green*/);
-        }
-        else
-        {
-            FormatColoredString(passFailMessage, "FAIL", 31 /*red*/);
-        }
-        Print::FormatToMiniUART("[{}] {}\r\n", passFailMessage, apMessage);
-    }
-
-    /**
-     * Tiny helper to emit a skip message
-     * 
-     * @param apMessage The message to emit
-    */
-    void EmitTestSkipResult(char const* const apMessage)
-    {
-        char skipMessage[32];
-        FormatColoredString(skipMessage, "SKIP", 33 /*yellow*/);
-        Print::FormatToMiniUART("[{}] {}\r\n", skipMessage, apMessage);
-    }
-
-    ///////////////////////////////////////////////////////////////////////////
-    // bitset tests
-    ///////////////////////////////////////////////////////////////////////////
-
-    /**
-     * Helper for testing operator &= for std::bitset
-     * 
-     * @param aInitialValue The initial value of the bitset
-     * @param aOtherValue The other value to &= with it
-     * @return The resulting bitset
-    */
-    template<class BitsetType>
-    constexpr auto BitsetAndAssign(unsigned long const aInitialValue, unsigned long const aOtherValue)
-    {
-        BitsetType result{ aInitialValue };
-        BitsetType other{ aOtherValue };
-        return result &= other;
-    }
-
-    /**
-     * Helper for testing operator & for std::bitset
-     * 
-     * @param aInitialValue The initial value of the bitset
-     * @param aOtherValue The other value to & with it
-     * @return The resulting bitset
-    */
-    template<class BitsetType>
-    constexpr auto BitsetAnd(unsigned long const aInitialValue, unsigned long const aOtherValue)
-    {
-        BitsetType result{ aInitialValue };
-        BitsetType other{ aOtherValue };
-        return result & other;
-    }
-
-    /**
-     * Helper for testing operator |= for std::bitset
-     * 
-     * @param aInitialValue The initial value of the bitset
-     * @param aOtherValue The other value to |= with it
-     * @return The resulting bitset
-    */
-    template<class BitsetType>
-    constexpr auto BitsetOrAssign(unsigned long const aInitialValue, unsigned long const aOtherValue)
-    {
-        BitsetType result{ aInitialValue };
-        BitsetType other{ aOtherValue };
-        return result |= other;
-    }
-
-    /**
-     * Helper for testing operator | for std::bitset
-     * 
-     * @param aInitialValue The initial value of the bitset
-     * @param aOtherValue The other value to | with it
-     * @return The resulting bitset
-    */
-    template<class BitsetType>
-    constexpr auto BitsetOr(unsigned long const aInitialValue, unsigned long const aOtherValue)
-    {
-        BitsetType result{ aInitialValue };
-        BitsetType other{ aOtherValue };
-        return result |= other;
-    }
-
-    /**
-     * Helper for testing operator ^= for std::bitset
-     * 
-     * @param aInitialValue The initial value of the bitset
-     * @param aOtherValue The other value to ^= with it
-     * @return The resulting bitset
-    */
-    template<class BitsetType>
-    constexpr auto BitsetXorAssign(unsigned long const aInitialValue, unsigned long const aOtherValue)
-    {
-        BitsetType result{ aInitialValue };
-        BitsetType other{ aOtherValue };
-        return result ^= other;
-    }
-
-    /**
-     * Helper for testing operator ^ for std::bitset
-     * 
-     * @param aInitialValue The initial value of the bitset
-     * @param aOtherValue The other value to ^ with it
-     * @return The resulting bitset
-    */
-    template<class BitsetType>
-    constexpr auto BitsetXor(unsigned long const aInitialValue, unsigned long const aOtherValue)
-    {
-        BitsetType result{ aInitialValue };
-        BitsetType other{ aOtherValue };
-        return result ^ other;
-    }
-
-    /**
-     * Helper for testing operator [] (set) for std::bitset
-     * 
-     * @param aInitialValue The initial value of the bitset
-     * @param aPos The bit to adjust
-     * @param aValue The value to give it
-     * @return The resulting bitset
-    */
-    template<class BitsetType>
-    constexpr auto BitsetIndexSet(unsigned long const aInitialValue, size_t const aPos, bool const aValue)
-    {
-        BitsetType result{ aInitialValue };
-        result[aPos] = aValue;
-        return result;
-    }
-
-    // An empty bitset
-    static_assert(sizeof(std::bitset<0>) == 1, "Unexpected size of empty bitset");
-    static_assert(std::bitset<0>{}.to_ulong() == 0UL, "Expected constructor for bitset to zero");
-    static_assert(std::bitset<0>{ 10 }.to_ulong() == 0UL, "Expected constructor for bitset to truncate");
-    static_assert(BitsetAndAssign<std::bitset<0>>(0b11, 0b101).to_ulong() == 0, "Unexpected &= result");
-    static_assert(BitsetOrAssign<std::bitset<0>>(0b11, 0b101).to_ulong() == 0, "Unexpected |= result");
-    static_assert(BitsetXorAssign<std::bitset<0>>(0b11, 0b101).to_ulong() == 0, "Unexpected ^= result");
-    // #TODO: Can't test set on bitset<0> because any bit is out of range
-    static_assert((~std::bitset<0>{ 0b11 }).to_ulong() == 0, "Unexpected ~ result");
-    static_assert(std::bitset<0>{ 0b11 }.flip().to_ulong() == 0, "Unexpected flip() result");
-    // #TODO: Can't test flip(pos) on bitset<0> because any bit is out of range
-    // We don't test operator[] on bitset<0> because it's UB when out of range (doesn't throw)
-    static_assert(std::bitset<0>{ 10 }.to_ullong() == 0ULL, "Unexpected to_ullong() result");
-    static_assert(std::bitset<0>{}.size() == 0, "Unexpected size() result");
-    // #TODO: Can't test test() on bitset<0> because any bit is out of range
-    static_assert(std::bitset<0>{ 10 }.all() == true, "Unexpected all() result");
-    static_assert(std::bitset<0>{ 10 }.any() == false, "Unexpected any() result");
-    static_assert(std::bitset<0>{ 10 }.none() == true, "Unexpected none() result");
-    static_assert(BitsetAnd<std::bitset<0>>(0b11, 0b101).to_ulong() == 0, "Unexpected & result");
-    static_assert(BitsetOr<std::bitset<0>>(0b11, 0b101).to_ulong() == 0, "Unexpected | result");
-    static_assert(BitsetXor<std::bitset<0>>(0b11, 0b101).to_ulong() == 0, "Unexpected ^ result");
-
-    // A bitset that is exactly the right size for its internal type
-    static_assert(sizeof(std::bitset<64>) == 8, "Unexpected size of full bitset");
-    static_assert(std::bitset<64>{}.to_ulong() == 0UL, "Expected constructor for bitset to zero");
-    static_assert(std::bitset<64>{ 10 }.to_ulong() == 10UL, "Expected constructor for bitset to set low bits");
-    static_assert(BitsetAndAssign<std::bitset<64>>(0b11, 0b101).to_ulong() == 0b001, "Unexpected &= result");
-    static_assert(BitsetOrAssign<std::bitset<64>>(0b11, 0b101).to_ulong() == 0b111, "Unexpected |= result");
-    static_assert(BitsetXorAssign<std::bitset<64>>(0b11, 0b101).to_ulong() == 0b110, "Unexpected ^= result");
-    static_assert(std::bitset<64>{ 0b11 }.set(2, true).to_ulong() == 0b111, "Unexpected set() result");
-    static_assert(std::bitset<64>{ 0b11 }.set(1, false).to_ulong() == 0b001, "Unexpected set() result");
-    static_assert(std::bitset<64>{ 0b11 }.reset(1).to_ulong() == 0b001, "Unexpected reset() result");
-    static_assert((~std::bitset<64>{ 0b11 }).to_ulong() == ~0b11UL, "Unexpected ~ result");
-    static_assert(std::bitset<64>{ 0b11 }.flip().to_ulong() == ~0b11UL, "Unexpected flip() result");
-    static_assert(std::bitset<64>{ 0b11 }.flip(1).to_ulong() == 0b01, "Unexpected flip(pos) result");
-    static_assert(std::bitset<64>{ 0b11 }.flip(2).to_ulong() == 0b111, "Unexpected flip(pos) result");
-    static_assert(std::bitset<64>{ 0b11 }[1] == true, "Unexpected operator[] result");
-    static_assert(BitsetIndexSet<std::bitset<64>>(0b11, 2, true).to_ulong() == 0b111, "Unexpected operator[] (set) result");
-    static_assert(BitsetIndexSet<std::bitset<64>>(0b11, 1, false).to_ulong() == 0b001, "Unexpected operator[] (set) result");
-    static_assert(std::bitset<64>{ 10 }.to_ullong() == 10ULL, "Unexpected to_ullong() result");
-    static_assert(std::bitset<64>{}.size() == 64, "Unexpected size() result");
-    static_assert(std::bitset<64>{ 0b11 }.test(1) == true, "Unexpected test() result");
-    static_assert(std::bitset<64>{ 0b11 }.test(2) == false, "Unexpected test() result");
-    static_assert(std::bitset<64>{ 10 }.all() == false, "Unexpected all() result");
-    static_assert(std::bitset<64>{ UINT64_MAX }.all() == true, "Unexpected all() result");
-    static_assert(std::bitset<64>{ 0 }.any() == false, "Unexpected any() result");
-    static_assert(std::bitset<64>{ 10 }.any() == true, "Unexpected any() result");
-    static_assert(std::bitset<64>{ 0 }.none() == true, "Unexpected none() result");
-    static_assert(std::bitset<64>{ 10 }.none() == false, "Unexpected none() result");
-    static_assert(BitsetAnd<std::bitset<64>>(0b11, 0b101).to_ulong() == 0b001, "Unexpected &= result");
-    static_assert(BitsetOr<std::bitset<64>>(0b11, 0b101).to_ulong() == 0b111, "Unexpected |= result");
-    static_assert(BitsetXor<std::bitset<64>>(0b11, 0b101).to_ulong() == 0b110, "Unexpected ^= result");
-    // #TODO: Find a good way to test swap on bool refs
-
-    // A bitset that is smaller than its internal type
-    static_assert(sizeof(std::bitset<32>) == 8, "Unexpected size of partial bitset");
-    static_assert(std::bitset<32>{}.to_ulong() == 0UL, "Expected constructor for bitset smaller than word size to zero");
-    static_assert(std::bitset<32>{ 10 }.to_ulong() == 10UL, "Expected constructor for bitset smaller than word size to set low bits");
-    static_assert(std::bitset<32>{ UINT64_MAX }.to_ulong() == UINT32_MAX, "Expected constructor for bitset smaller than word size to truncate large input");
-    static_assert(BitsetAndAssign<std::bitset<32>>(0b11, 0b101).to_ulong() == 0b001, "Unexpected &= result");
-    static_assert(BitsetOrAssign<std::bitset<32>>(0b11, 0b101).to_ulong() == 0b111, "Unexpected |= result");
-    static_assert(BitsetXorAssign<std::bitset<32>>(0b11, 0b101).to_ulong() == 0b110, "Unexpected ^= result");
-    static_assert(std::bitset<32>{ 0b11 }.set(2, true).to_ulong() == 0b111, "Unexpected set() result");
-    static_assert(std::bitset<32>{ 0b11 }.set(1, false).to_ulong() == 0b001, "Unexpected set() result");
-    static_assert(std::bitset<32>{ 0b11 }.reset(1).to_ulong() == 0b001, "Unexpected reset() result");
-    static_assert((~std::bitset<32>{ 0b11 }).to_ulong() == ~0b11U, "Unexpected ~ result");
-    static_assert(std::bitset<32>{ 0b11 }.flip().to_ulong() == ~0b11U, "Unexpected flip() result");
-    static_assert(std::bitset<32>{ 0b11 }.flip(1).to_ulong() == 0b01, "Unexpected flip(pos) result");
-    static_assert(std::bitset<32>{ 0b11 }.flip(2).to_ulong() == 0b111, "Unexpected flip(pos) result");
-    static_assert(std::bitset<32>{ 0b11 }[1] == true, "Unexpected operator[] result");
-    static_assert(BitsetIndexSet<std::bitset<32>>(0b11, 2, true).to_ulong() == 0b111, "Unexpected operator[] (set) result");
-    static_assert(BitsetIndexSet<std::bitset<32>>(0b11, 1, false).to_ulong() == 0b001, "Unexpected operator[] (set) result");
-    static_assert(std::bitset<32>{ 10 }.to_ullong() == 10ULL, "Unexpected to_ullong() result");
-    static_assert(std::bitset<32>{}.size() == 32, "Unexpected size() result");
-    static_assert(std::bitset<32>{ 0b11 }.test(1) == true, "Unexpected test() result");
-    static_assert(std::bitset<32>{ 0b11 }.test(2) == false, "Unexpected test() result");
-    static_assert(std::bitset<32>{ 10 }.all() == false, "Unexpected all() result");
-    static_assert(std::bitset<32>{ UINT32_MAX }.all() == true, "Unexpected all() result"); // makes sure that the top 32 bits don't count
-    static_assert(std::bitset<32>{ 0 }.any() == false, "Unexpected any() result");
-    static_assert(std::bitset<32>{ 10 }.any() == true, "Unexpected any() result");
-    static_assert(std::bitset<32>{ 0 }.none() == true, "Unexpected none() result");
-    static_assert(std::bitset<32>{ 10 }.none() == false, "Unexpected none() result");
-    static_assert(BitsetAnd<std::bitset<32>>(0b11, 0b101).to_ulong() == 0b001, "Unexpected &= result");
-    static_assert(BitsetOr<std::bitset<32>>(0b11, 0b101).to_ulong() == 0b111, "Unexpected |= result");
-    static_assert(BitsetXor<std::bitset<32>>(0b11, 0b101).to_ulong() == 0b110, "Unexpected ^= result");
-    // #TODO: Find a good way to test swap on bool refs
-
-    // A bitset that is larger than its internal type, partially filled
-    static_assert(sizeof(std::bitset<96>) == 16, "Unexpected size of large bitset");
-    static_assert(std::bitset<96>{}.to_ulong() == 0UL, "Expected constructor for bitset larger than word size to zero");
-    static_assert(std::bitset<96>{ 10 }.to_ulong() == 10UL, "Expected constructor for bitset larger than word size to set low bits");
-    // #TODO: Need a good way to test high numbers with large bitsets, since to_ulong won't return the right value for
-    // any bits over 64 (cause it's returning an unsigned long). And when we put in the range checking it'll throw
-    // anyway
-    static_assert(std::bitset<96>{ 0 }.set(65, true)[65] == true, "Unexpected set() result");
-    static_assert(std::bitset<96>{ 0 }.flip(65)[65] == true, "Unexpected flip(pos) result");
-    // operator[] (get) being tested above
-    static_assert(BitsetIndexSet<std::bitset<96>>(0, 65, true)[65] == true, "Unexpected operator[] (set) result");
-    // to_ulong tested above
-    static_assert(std::bitset<96>{ 10 }.to_ullong() == 10ULL, "Unexpected to_ullong result");
-    static_assert(std::bitset<96>{}.size() == 96, "Unexpected size() result");
-    static_assert(std::bitset<96>{ 0 }.set(65).test(65) == true, "Unexpected test() result");
-    static_assert(std::bitset<96>{ 0 }.set(65).test(66) == false, "Unexpected test() result");
-    static_assert(std::bitset<96>{ 10 }.all() == false, "Unexpected all() result");
-    static_assert(std::bitset<96>{ UINT64_MAX }.all() == false, "Unexpected all() result");
-    static_assert(std::bitset<96>{ 0 }.flip().all() == true, "Unexpected all() result");
-    static_assert(std::bitset<96>{ 0 }.any() == false, "Unexpected any() result");
-    static_assert(std::bitset<96>{ 0 }.set(65, true).any() == true, "Unexpected any() result");
-    static_assert(std::bitset<96>{ 0 }.none() == true, "Unexpected none() result");
-    static_assert(std::bitset<96>{ 10 }.none() == false, "Unexpected none() result");
-    // #TODO: Find a good way to test swap on bool refs
+    using namespace ::UnitTests;
 
     ///////////////////////////////////////////////////////////////////////////
     // climits/limits.h tests
@@ -906,6 +635,8 @@ namespace UnitTests
         ExceptionLevelTest();
         FloatingPointTest();
         SIMDTest();
+
+        KernelStdlib::Bitset::Run();
         
         StdMoveTest();
         StdForwardTest();
