+++ kernel/UnitTests/CMakeLists.txt
@@ -1,0 +1,6 @@
+target_sources(kernel8.elf
+    PRIVATE
+        Framework.h Framework.cpp
+)
+
+add_subdirectory(KernelStdlib)
