+++ kernel/UnitTests/KernelStdlib/BitsetTests.h
@@ -1,0 +1,12 @@
+#ifndef KERNEL_UNITTESTS_KERNELSTDLIB_BITSETTESTS_H
+#define KERNEL_UNITTESTS_KERNELSTDLIB_BITSETTESTS_H
+
+namespace UnitTests::KernelStdlib::Bitset
+{
+    /**
+     * Run all runtime tests
+     */
+    void Run();
+}
+
+#endif // KERNEL_UNITTESTS_KERNELSTDLIB_BITSETTESTS_H
