+++ kernel/UnitTests/KernelStdlib/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(kernel8.elf
+    PRIVATE
+        BitsetTests.h BitsetTests.cpp
+)
